// projects/recipes/src/app/models/enum/sort.enum.ts
export enum SortType {
  SORT_SEARCH_RESULTS = "SORT_SEARCH_RESULTS",
  SORT_CUSTOM = "SORT_CUTOM",
  SORT_ALPHABET = "SORT_ALPHABET",
  SORT_DIFFICULTY = "SORT_DIFFICULTY",
  SORT_RATING = "SORT_RATING",
  SORT_PREPARED_QUANTITY = "SORT_PREPARED_QUANTITY",
  SORT_LAST_PREPARED_DATE = "SORT_LAST_PREPARED_DATE",
  SORT_LAST_EDITED_DATE = "SORT_LAST_EDITED_DATE",
  SORT_CREATED_DATE = "SORT_CREATED_DATE",
  SORT_PREPARATION_TIME = "SORT_PREPARATION_TIME",
  SORT_INGREDIENTS_QUANTITY = "SORT_INGREDIENTS_QUANTITY",
  SORT_PLANNED = "SORT_PLANNED",
  SORT_COSTS = "SORT_COSTS",
  SORT_PORTIONS_LEFT = "SORT_PORTIONS_LEFT",
  SORT_USE_UNTIL = "SORT_USE_UNTIL",
  SORT_AVAILABLE = "SORT_AVAILABLE",
}

export enum SortKey {
  SORT_NONE = "none",
  SORT_SEARCH_RESULTS = "search-results",
  SORT_CUSTOM = "custom",
  SORT_ALPHABET = "alphabet",
  SORT_DIFFICULTY = "difficulty",
  SORT_RATING = "rating",
  SORT_PREPARED_QUANTITY = "prepared-quantity",
  SORT_LAST_PREPARED_DATE = "last-prepared-date",
  SORT_LAST_EDITED_DATE = "last-edited-date",
  SORT_CREATED_DATE = "created-date",
  SORT_PREPARATION_TIME = "preparation-time",
  SORT_INGREDIENT_QUANTITY = "ingredient-quantity",
  SORT_PLANNED = "planned",
  SORT_COSTS = "costs",
  SORT_PORTIONS_LEFT = "portions-left",
  SORT_USE_UNTIL = "use-until",
  SORT_AVAILABLE = "available",
}